Refact: separate component files
